Update 게시글 수정 API
①불필요한 파라미터 제거
 - postImageIdxList
②형식적 Validation 코드 추가
 - 게시글 내용, 이미지, 이미지 번호 미입력 여부 확인
 - 이미지 번호 입력시 이미지 파일도 입력 되었는지 확인
 - 이미지 파일 입력시 이미지 번호도 입력 되었는지 확인
③의미적 Validation 코드 추가
 - 게시글 삭제 여부 확인
④게시글 내용과 이미지,(+이미지 번호) 중 하나만 입력해도 게시글 정보가 변경되도록 로직 구현 : if문 활용
⑤이미지 정보 변경 로직 수정
 - 기존 : post_image 테이블에서 postIdx에 해당되는 레코드의 이미지(+이미지 번호) 정보를 Update
   //이 과정에서 문제 발생 : 기존에 post_image 테이블에 등록한 이미지 개수와 입력받은 이미지의 개수가 일치해야 함
 - 변경 후 : post_image 테이블에서 postIdx에 해당되는 레코드를 Delete 후 새로 입력받은 이미지(+이미지 번호)를 post_image 테이블에 Insert

+++ src/main/java/com/instagram/domain/post/service/PostService.java
@@ -10,7 +10,6 @@ import com.instagram.domain.post.dto.GetPostsRes;
 import com.instagram.domain.user.dao.UserDao;
 import com.instagram.domain.user.domain.User;
 import com.instagram.global.error.BasicException;
-import com.instagram.global.error.BasicResponse;
 import com.instagram.global.util.Security.Secret;
 import com.instagram.infra.aws.AwsS3Service;
 import lombok.RequiredArgsConstructor;
@@ -18,7 +17,6 @@ import org.springframework.beans.BeanUtils;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.net.URI;
@@ -114,10 +112,15 @@ public class PostService {
 
     /* 8. 게시글 수정 API */
     @Transactional(rollbackFor = {Exception.class})
-    public String modifyPost(Long postIdx, String content, List<Long> postImageIdxList, List<MultipartFile> multipartFile, List<Integer> imageNumber, Long userIdx) throws BasicException {
+    public String modifyPost(Long postIdx, String content, List<MultipartFile> multipartFile, List<Integer> imageNumberlist, Long userIdx) throws BasicException {
 
         Post postBefore = postDao.findByIdx(postIdx);
         List<PostImage> PostImageBeforeList = postImageDao.findByPost(postBefore);
+
+        //게시글 삭제 여부 조회
+        if(postBefore == null){             //게시글이 삭제되었다면..
+            throw new BasicException(RES_ERROR_POSTS_DELETE_POST);    //"삭제된 게시글"
+        }
 
         //게시글 작성자 일치 여부 확인
         User Writer = userDao.findByIdx(userIdx);
@@ -125,37 +128,61 @@ public class PostService {
             throw new BasicException(RES_ERROR_POSTS_NOT_SAME_POST);    //"개시글 작성자 불일치 오류"
         }
 
-        //Post DB에서 게시글 정보 수정
-        postDao.modifyPost(content, postIdx);
+        //게시글 내용 변경
+        if(content != null) {
+            try {
+                postDao.modifyPost(content, postIdx);
+            } catch (Exception exception) {
+                throw new BasicException(DATABASE_ERROR_MODIFY_FAIL_POSTS_CONTENT);   //"게시글 내용 변경 오류"
+            }
+        }
 
-        //PostImage DB에서 게시글 이미지 정보 수정
+        //게시글 이미지 정보 변경
         List<String> UUID_fileNameList = new ArrayList<>();;
+        List<PostImage> postImageListCreation = new ArrayList<>();
 
-        for(int i=0; i<postImageIdxList.size(); i++) {
-            UUID_fileNameList.add(awsS3Service.createFileNameToDB(multipartFile.get(i)));         //UUID가 적용된 새로운 이미지 파일명 리턴
-            postImageDao.modifyPostImage(Secret.AWS_S3_CONNECT_URL+UUID_fileNameList.get(i), postImageIdxList.get(i), postBefore);
-        }
+        if(imageNumberlist != null) {
+            try {
+                //DB에서 기존 이미지 삭제
+                postImageDao.deletePostImage(postBefore);
+
+                //DB에 새로운 이미지 등록
+                for(int i=0; i<imageNumberlist.size(); i++) {
+                    PostImage postImageCreation = new PostImage();   //내부에서 postImageIdx를 반복 횟수만큼 생성 (for문에 꼭 포함해야 한다)
+                    postImageCreation.setPost(postBefore);
+                    postImageCreation.setImageNumber(imageNumberlist.get(i));
+
+                    UUID_fileNameList.add(awsS3Service.createFileNameToDB(multipartFile.get(i)));         //UUID가 적용된 새로운 이미지 파일명 리턴
+                    postImageCreation.setImage(Secret.AWS_S3_CONNECT_URL+UUID_fileNameList.get(i));
+
+                    postImageListCreation.add(postImageCreation);
+                }
+                postImageDao.saveAll(postImageListCreation);
+            }
+            catch (Exception exception) {
+                throw new BasicException(DATABASE_ERROR_MODIFY_FAIL_POSTS_IMAGE);   //"게시글 이미지 변경 오류"
+            }
 
 
-        //S3에 새로운 이미지 파일 업로드
-        for(int i=0; i<postImageIdxList.size(); i++) {
-            URI imageUrl = awsS3Service.uploadFile(multipartFile.get(i), UUID_fileNameList.get(i));  //이미지 파일과 UUID가 적용된 파일명 인수로 전달
-        }
+            //S3에 새로운 이미지 파일 업로드
+            for(int i=0; i<imageNumberlist.size(); i++) {
+                URI imageUrl = awsS3Service.uploadFile(multipartFile.get(i), UUID_fileNameList.get(i));  //이미지 파일과 UUID가 적용된 파일명 인수로 전달
+            }
 
-        //S3에서 변경 전 이미지 파일 삭제
-        try {
-            for(int i=0; i<postImageIdxList.size(); i++) {
-                String beforeS3_fileName = PostImageBeforeList.get(i).getImage().replace(Secret.AWS_S3_CONNECT_URL, "");
-                awsS3Service.deleteFile(beforeS3_fileName);
+            //S3에서 변경 전 이미지 파일 삭제
+            try {
+                for(int i=0; i<PostImageBeforeList.size(); i++) {
+                    String beforeS3_fileName = PostImageBeforeList.get(i).getImage().replace(Secret.AWS_S3_CONNECT_URL, "");
+                    awsS3Service.deleteFile(beforeS3_fileName);
+                }
+            }
+            catch(Exception exception){    //이전 이미지 파일 삭제 실패시 업로드된 새로운 파일을 삭제
+                for(int i=0; i<imageNumberlist.size(); i++) {
+                    awsS3Service.deleteFile(UUID_fileNameList.get(i));
+                }
+                throw new BasicException(S3_ERROR_FAIL_DELETE_FILE);  //"S3에서 파일 삭제 실패"
             }
         }
-        catch(Exception exception){    //이전 이미지 파일 삭제 실패시 업로드된 새로운 파일을 삭제
-            for(int i=0; i<postImageIdxList.size(); i++) {
-                awsS3Service.deleteFile(UUID_fileNameList.get(i));
-            }
-            throw new BasicException(S3_ERROR_FAIL_DELETE_FILE);  //"S3에서 파일 삭제 실패"
-        }
-
         return "게시글 수정 성공";
     }
